+++ CHANGELOG.md
@@ -6,6 +6,23 @@
 > make sure you follow our [migration guide](https://docs.sentry.io/platforms/react-native/migration/) first.
 <!-- prettier-ignore-end -->
 
+## Unreleased
+
+### Features
+
+- Add `wrapExpoImage` and `wrapExpoAsset` for Expo performance monitoring ([#5427](https://github.com/getsentry/sentry-react-native/issues/5427))
+  - `wrapExpoImage` instruments `Image.prefetch` and `Image.loadAsync` from `expo-image`
+  - `wrapExpoAsset` instruments `Asset.loadAsync` from `expo-asset`
+  ```js
+  import { Image } from 'expo-image';
+  import { Asset } from 'expo-asset';
+  import * as Sentry from '@sentry/react-native';
+
+  Sentry.wrapExpoImage(Image);
+  Sentry.wrapExpoAsset(Asset);
+  ```
+- Adds tags with Expo Updates context variables to make them searchable and filterable ([#5788](https://github.com/getsentry/sentry-react-native/pull/5788))
+
 ## 8.3.0
 
 ### Features